+++ NEWS.md
@@ -1,3 +1,7 @@
+# RNAseqQC 0.1.2
+* Fixed issue when the results object fed into plot_ma() contains s-values.
+* Add GC content gene annotation. 
+
 # RNAseqQC 0.1.1
 * Fixed some package dependencies by moving ggsci to Imports and updating the data vignette.